+++ packages/xcfreader/src/examples/text.ts
@@ -1,5 +1,5 @@
 import { Logger } from "../lib/logger.js";
-import { XCFParser as GimpParser, XCFImage } from "../gimpparser.js";
+import { XCFParser as GimpParser, XCFPNGImage } from "../gimpparser.js";
 import fs from "fs";
 import path from "path";
 import { fileURLToPath } from "url";
@@ -15,10 +15,11 @@ if (!fs.existsSync(outDir)) fs.mkdirSync(outDir, { recursive: true });
   try {
     const parser = await GimpParser.parseFileAsync(xcfPath);
     const layers = parser.layers;
-    const image = new XCFImage(parser.width, parser.height);
+    const image = new XCFPNGImage(parser.width, parser.height);
 
     for (const layer of layers.slice().reverse()) {
-      const layerImage = layer.makeImage();
+      const layerImage = new XCFPNGImage(layer.width, layer.height);
+      layer.makeImage(layerImage);
       layer.makeImage(image, true);
       await layerImage.writeImage(path.resolve(outDir, layer.name + ".png"));
       Logger.log(layer.parasites);
@@ -27,7 +28,9 @@ if (!fs.existsSync(outDir)) fs.mkdirSync(outDir, { recursive: true });
     await image.writeImage(path.resolve(outDir, "multi1.png"));
     Logger.log("Image 1 saved");
 
-    await parser.createImage().writeImage(path.resolve(outDir, "multi2.png"));
+    const image2 = new XCFPNGImage(parser.width, parser.height);
+    parser.createImage(image2);
+    await image2.writeImage(path.resolve(outDir, "multi2.png"));
     Logger.log("Image 2 saved");
   } catch (err) {
     Logger.error(err);